перенёс методы switch_to_new_tab и is_element_present в файл base_page.py
